--- controllers/animals.js
@@ -1,50 +1,0 @@
-// Delete this and actually add Animals, the Survey will be on the HTML
-// our animals are the Mr/T pitbull, The polar bear and Black Panther
-var Animals = [
-    {
-        name: "T'challa, the Black Panther",
-        photo: "../images/Black_Panther.jpeg",
-        scores: [
-         "Cat",
-         "Panther",
-         "Dog",
-         "* In Soulja Boy voice* SWEATERS",
-         "cheetah",
-         "yes"
-
-        ]
-
-    },
-   
-    {
-        name: "I pitbull the Fool",
-        photo:"../images/I_Pitbull_the_Fool.jpg",
-        scores: [
-            "Pigeon", 
-            "Eagle",
-            "Panda",
-            "False",
-            "Alligator",
-            "Who's Arthur"
-        ]
-    },
-
-    {
-        name: "Polar Sweater",
-        photo:"../images/Polar_Sweater.jpg",
-        scores:[
-            "Dolphin",
-            "Shark",
-            "Penguin",
-            "True",
-            "Koala",
-            "No"
-        ]
-    }
-]
-
-
-
-module.exports = animals;
-
-
