+++ public/index.html
@@ -81,7 +81,7 @@
       <div class="parallax-container valign-wrapper">
             <div class="section no-pad-bot">
               <div class="container">
-                <div class="row center">
+                <div class="row center" id="fantasy">
                   <h5 class="header col s12 light">A Way to Live Out a Childhood Fantasy</h5>
                 </div>
               </div>
@@ -91,7 +91,7 @@
 
 
     <!--JavaScript at end of body for optimized loading-->
-    <script src="https://code.jquery.com/jquery-3.3.1.js" integrity="sha256-2Kok7MbOyxpgUVvAk/HJ2jigOSYS2auK4Pfzbm7uH60="
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/core.js" integrity="sha256-2Kok7MbOyxpgUVvAk/HJ2jigOSYS2auK4Pfzbm7uH60="
         crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
     <script src="js/logic.js"></script>
