+++ public/survey.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
 
     <!-- Our CSS -->
-    <link href="https://fonts.googleapis.com/css?family=Bungee+Shade" rel="stylesheet">
+    <!-- <link href="https://fonts.googleapis.com/css?family=Bungee+Shade" rel="stylesheet"> -->
     <link rel="stylesheet" href="css/style.css">
 
     <!--Let browser know website is optimized for mobile-->
@@ -37,7 +37,7 @@
                         <p>If any of these animals could talk which would be the rudest?</p>
                     </div>
                     <div class="card-action">
-                        <select class="choices" id="q1">
+                        <select class="chosen-select" id="q1">
                             <option value="" disabled selected>Choose your option</option>
                             <option value="1">Pigeon</option>
                             <option value="2">Cat</option>
@@ -56,7 +56,7 @@
                         
                     </div>
                     <div class="card-action">
-                        <select class="choices" id="q2">
+                        <select class="chosen-select" id="q2">
                             <option value="" disabled selected>Choose your option</option>
                             <option value="1">Eagle</option>
                             <option value="2">Panther</option>
@@ -78,11 +78,11 @@
                         <p>If any of these animals could talk, who would be the nicest?</p>
                     </div>
                     <div class="card-action">
-                        <select class="choices" id="q3">
+                        <select class="chosen-select" id="q3">
                             <option value="" disabled selected>Choose your option</option>
-                            <option>Panda</option>
-                            <option>Dog</option>
-                            <option>Penguin</option>
+                            <option value="1"> Panda</option>
+                            <option value="2">Dog</option>
+                            <option value="3">Penguin</option>
                         </select>
                         <label>Answers</label>
                     </div>
@@ -96,11 +96,11 @@
                         <br>
                     </div>
                     <div class="card-action">
-                        <select class="choices" id="q4">
+                        <select class="chosen-select" id="q4">
                             <option value="" disabled selected>Choose your option</option>
-                            <option>True</option>
-                            <option>False</option>
-                            <option>*In Soulja Boy voice* SWEATERS</option>
+                            <option value="1"> True</option>
+                            <option value="2">False</option>
+                            <option value="3">*In Soulja Boy voice* SWEATERS</option>
                         </select>
                         <label>Answers</label>
                     </div>
@@ -118,11 +118,11 @@
                         <p>Which Would You prefer as a pet?</p>
                     </div>
                     <div class="card-action">
-                        <select class="choices" id="q5">
+                        <select class="chosen-select"id="q5">
                             <option value="" disabled selected>Choose your option</option>
-                            <option>Alligator</option>
-                            <option>Cheetah</option>
-                            <option>Koala</option>
+                            <option value= "1">Alligator</option>
+                            <option value="2">Cheetah</option>
+                            <option value="3">Koala</option>
                         </select>
                         <label>Answers</label>
                     </div>
@@ -135,11 +135,11 @@
                         <p> Did You know Arthur was and is an Aardvark?</p>
                     </div>
                     <div class="card-action">
-                        <select class="choices" id="q6">
+                        <select class="chosen-select" id="q6">
                             <option value="" disabled selected>Choose your option</option>
-                            <option>Yes</option>
-                            <option>No</option>
-                            <option>Who's Arthur</option>
+                            <option value="1">Yes</option>
+                            <option value="2">No</option>
+                            <option value="3">Who's Arthur</option>
                         </select>
                         <label>Answers</label>
                     </div>
